# src/openapi.json
{
  "openapi": "2.8.0",
  "info": {
    "title": "XTE Apis",
    "author": "D2:业",
    "version": "v1.0.0",
    "description": "Simple and easy to use API."
  },
  "servers": [
    {
      "url": "https://api.xte.web.id"
    }
  ],
...
  "paths": {
...
    "/v1/maker/toemas": {
      "get": {
        "summary": "AI Golden Effect",
        "description": "Mengubah subjek menjadi patung emas mewah. Output: Image Buffer.",
        "tags": ["Maker"],
        "parameters": [{ "name": "url", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Image JPEG" } }
      }
    },
...
    "/v1/stalk/ig": {
      "get": {
        "summary": "Instagram Profile Stalker",
        "description": "Melihat informasi profil publik pengguna Instagram.",
        "tags": ["Stalk"],
        "parameters": [{ "name": "username", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Success" } }
      }
    },
...
    "/tools/bahlil": {
      "get": {
        "summary": "Brat meme bahlil",
        "description": "Membuat meme menggunakan gambar Bahlil dengan teks kustom.",
        "tags": ["Tools"],
        "parameters": [{ "name": "text", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Image PNG" } }
      }
    },
...
    "/v1/tools/chord": {
      "get": {
        "summary": "Guitar Chord & Lyrics Finder",
        "description": "Cari chord gitar dan lirik lagu favoritmu.",
        "tags": ["Tools"],
        "parameters": [{ "name": "q", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Success" } }
      }
    },
...
    "/tools/undress": {
      "get": {
        "summary": "AI Image Alter",
        "description": "Alat manipulasi gambar (dibatasi/restricted dengan token khusus).",
        "tags": ["Tools"],
        "parameters": [
          { "name": "url", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "token", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Success" } }
      }
    },
...
    "/v1/search/pinterest": {
      "get": {
        "summary": "Pinterest Image Search",
        "description": "Mencari koleksi gambar di platform Pinterest.",
        "tags": ["Search"],
        "parameters": [{ "name": "q", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Success" } }
      }
    },
...
    "/v1/search/soundcloud": {
      "get": {
        "summary": "SoundCloud Search",
        "description": "Mencari lagu di SoundCloud secara real-time. Menampilkan judul, artis, link track, dan thumbnail HD.",
        "tags": ["Search"],
        "parameters": [{ "name": "q", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Success" } }
      }
    },
...
  }
}